Move routing to views.py
