+++ api/requirements.txt
@@ -13,3 +13,4 @@ Shapely==1.7.1
 geoalchemy2==0.8.5
 pillow==7.2.0
 pytest==6.2.4
+imagehash==4.2.0
