+++ api/src/everything/__init__.py
@@ -1,0 +1,6 @@
+import os
+from flask import Flask
+
+app = Flask(__name__)
+
+import everything.app
