+++ api/src/everything/media_io.py
@@ -1,7 +1,13 @@
-from datetime import datetime
+import os
 import sys
 import subprocess
+from datetime import datetime
 from everything.db import Media
+
+
+image_extensions = ['.jpg', '.png', '.jpeg']
+video_extensions = ['.mp4', '.mov']
+allowed_extensions = image_extensions + video_extensions
 
 
 def _format_date_to_touch(datetime: datetime) -> str:
@@ -34,3 +40,41 @@ def write_date_to_media(media: Media):
 
 	# Run command
 	subprocess.run(command, shell=True, check=True)
+
+
+def get_extension(filepath: str):
+	"""Returns the extension of a filepath
+	
+	Parameters:
+		filepath (str): The filepath to retreive the extension of
+	"""
+	_, ext = os.path.splitext(filepath)
+	return ext
+
+
+def is_image(filepath: str) -> bool:
+	"""Returns whether a filepath is an image"""
+	ext = get_extension(filepath)
+	return ext.lower() in image_extensions
+
+
+def is_media_file(filepath: str) -> bool:
+	"""Returns whether this file can be stored as Media
+	
+	Parameters:
+		filepath (str): Full filepath of file
+	"""
+	ext = get_extension(filepath)
+	return ext.lower() in allowed_extensions
+
+
+def mime_from_ext(filepath: str):
+	"""TODO"""
+	ext = get_extension(filepath)
+	return {
+		'.jpg': 'image/jpeg',
+		'.jpeg': 'image/jpeg',
+		'.png': 'image/png',
+		'.mp4': 'video/mp4',
+		'.mov': 'application/octet-stream'
+	}[ext.lower()]
